moved create_subsampling_error_grid function to experiments.py
