+++ scripts/experiments/2020-11-05-simple-mfbo.py
@@ -11,15 +11,13 @@ from warnings import warn
 import numpy as np
 import pandas as pd
 import xarray as xr
-from itertools import product
 from collections import namedtuple
 from operator import itemgetter
 from scipy.optimize import minimize
 from sklearn.linear_model import LinearRegression
 from pyprojroot import here
 
-from experiments import bi_fidelity_doe, scale_to_function, split_bi_fidelity_doe, \
-    BiFidelityDoE, mlcs
+from experiments import bi_fidelity_doe, scale_to_function, create_subsampling_error_grid, mlcs
 
 
 save_dir = here('files/2020-11-05-simple-mfbo/')
@@ -40,77 +38,6 @@ def fit_lin_reg(da: xr.DataArray, calc_SSE: bool=False):
     pred_y = reg.predict(X)
     SSE = np.sum((pred_y - y)**2)
     return reg, SSE
-
-
-def create_subsampling_error_grid(
-        archive: mlcs.CandidateArchive,
-        num_reps: int=50,
-        interval: int=2,
-        func=None
-) -> xr.DataArray:
-    """Create an error grid through subsampling from the given archive
-
-    :param archive:  `mlcs.CandidateArchive` containing all available evaluated candidates
-    :param num_reps: Number of independent repetitions for each size
-    :param interval: Interval at which to fill the error grid
-    :param func:     Multi-fidelity function to re-evaluate known candidates  #TODO: remove argument
-    :return:         `xr.DataArray` ErrorGrid
-    """
-
-    if not func:
-        # TODO: implement that the original function is not re-evaluated
-        # but that values from `archive` are re-used.
-        raise NotImplementedError
-
-    highs = archive.getcandidates(fidelity='high').candidates
-    lows = archive.getcandidates(fidelity='low').candidates
-    DoE = BiFidelityDoE(highs, lows)
-
-    max_num_high = len(highs)
-    max_num_low = len(lows)
-
-    error_grid = np.full((max_num_high+1, max_num_low+1, num_reps+1, 3), np.nan)
-
-    instances = [
-        (h, l, r)
-        for h, l, r in product(range(2, max_num_high, interval),
-                               range(3, max_num_low, interval),
-                               range(num_reps))
-        if h < l
-    ]
-
-    for i, (num_high, num_low, rep) in enumerate(instances):
-
-        if i % 1_000 == 0:
-            print(f"{i}/{len(instances)}")
-
-        # Create sub-sampled Multi-Fidelity DoE in- and output according to instance specification
-        (high_x, low_x), _ = split_bi_fidelity_doe(DoE, num_high, num_low)
-        high_y, low_y = func.high(high_x), \
-                        func.low(low_x)
-
-        # Create an archive from the MF-function and MF-DoE data
-        archive = mlcs.CandidateArchive.from_multi_fidelity_function(func, ndim=func.ndim)
-        archive.addcandidates(low_x, low_y, fidelity='low')
-        archive.addcandidates(high_x, high_y, fidelity='high')
-
-        # (Automatically) Create the hierarchical model
-        mfbo = mlcs.MultiFidelityBO(func, archive)
-
-        # Get the results we're interested in from the model for this instance
-        error_grid[num_high, num_low, rep] = mfbo.getMSE()
-
-    models = ['high_hier', 'high', 'low']
-    return xr.DataArray(
-        error_grid,
-        dims=['n_high', 'n_low', 'rep', 'model'],
-        coords={
-            'n_high': range(max_num_high+1),
-            'n_low': range(max_num_low+1),
-            'rep': range(num_reps+1),
-            'model': models,
-        },
-    )
 
 
 def simple_multifid_bo(func, budget, cost_ratio, doe_n_high, doe_n_low, num_reps=50):
